fix: Copy cozy-bar v7 assets only if should add public config
